+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2020-12-04
 ### Added
 - Add `coverage` command
   - It outputs the coverage after running `batler test`. For more information, run `batler coverage --help`.
@@ -23,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Support for running tests using any available simulator
 
 
-[Unreleased]: https://github.com/vitorbaraujo/batler/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/vitorbaraujo/batler/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/vitorbaraujo/batler/releases/tag/0.3.0
 [0.2.0]: https://github.com/vitorbaraujo/batler/releases/tag/0.2.0
 [0.1.0]: https://github.com/vitorbaraujo/batler/releases/tag/0.1.0